Split board auto size tests into individual files
